// src/enums/draw-type.ts
export enum DrawType {
    FILL_ROTATED_QUAD = 1,
    DRAW_ROTATED_QUAD = 2,
    FILL_LINED_ROTATED_QUAD = 3,

    FILL_AA_QUAD = 4,
    DRAW_AA_QUAD = 5,
    FILL_LINED_AA_QUAD = 6,

    FILL_ELLIPSOID = 7,
    DRAW_ELLIPSOID = 8,
    FILL_LINED_ELLIPSOID = 9,

    FILL_CIRCLE = 10,
    DRAW_CIRCLE = 11,
    FILL_LINED_CIRCLE = 12,

    DRAW_ELLIPSOIDAL_ARC = 13,
    DRAW_CIRCULAR_ARC = 14,

    DRAW_LINE = 15,
    DRAW_POLY_LINE = 16,

    FILL_RING_SLICE = 17,
    DRAW_RING_SLICE = 18,
    FILL_LINED_RING_SLICE = 19,

    FILL_CIRCLE_SLICE = 20,
    DRAW_CIRCLE_SLICE = 21,
    FILL_LINED_CIRCLE_SLICE = 22,
}